requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
appnope==0.1.2
async-timeout==4.0.0
attrs==21.2.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
dateparser==1.1.0
decorator==5.1.0
frozenlist==1.2.0
idna==3.3
ipython==7.28.0
jedi==0.18.0
lxml==4.6.3
matplotlib-inline==0.1.3
multidict==5.2.0
multitasking==0.0.9
numpy==1.21.3
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.21
ptyprocess==0.7.0
py==1.10.0
Pygments==2.10.0
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2021.11.1
requests==2.26.0
retry==0.9.2
simplejson==3.17.5
six==1.16.0
TA-Lib==0.4.21
traitlets==5.1.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
urllib3==1.26.7
wcwidth==0.2.5
websockets==9.1
yarl==1.7.2
